refactor(adapter-codex): move ThreadItem payload structs to thread_item_variants
item_types.rs was approaching the 300-line ceiling before B1 adds seven more
variants. Pure move, no behavior change — existing crate::item_types::X call
sites keep compiling via a glob re-export.
